Отделили 'FilmData' от 'FilmDetails'.
